+++ src/components/index.ts
@@ -1,3 +1,5 @@
-export * from './Sample'
+export * from './Badge'
 
-export * from './Badge';
+export * from './Button'
+
+export * from './Typography'