+++ src/libs/animation/variant/Slide/index.ts
@@ -1,3 +1,3 @@
-export type { SlideAnimationProps as SlideProps } from './type'
+export type { SlideProps } from './type'
 
 export { genAnimation as genSlideAnimation } from './gen'