+++ src/libs/animation/variant/Squeeze/type.ts
@@ -1,10 +1,10 @@
 import type { AnimationBaseProps } from '../type'
 
 export type OptionProps = {
-  direction?: 'left' | 'right' | 'top' | 'bottom'
+  axis?: 'x' | 'y'
 }
 
-export type SlideProps = {
-  key: 'slide'
+export type SqueezeProps = {
+  key: 'squeeze'
   option: OptionProps & AnimationBaseProps
 }